Moved timestamps aligning code to separate module
